+++ changelog.md
@@ -5,6 +5,8 @@ sortOrder: 4
 ---
 ## Pending `v0.3.13`
 - Add ability for `bencher run` to read output from a file without running a command argument
+- Improve Report creation error messages
+- Add filtering Reports by Branch, Testbed, and start/end time
 
 ## `v0.3.12`
 - Change Metric `lower_bound` and `upper_bound` to `lower_value` and `upper_value` respectively